refactor braket qasm (#355)
