+++ qbraid/transpiler/cirq_braket/__init__.py
@@ -20,10 +20,7 @@ Braket Conversions  (:mod:`qbraid.transpiler.cirq_braket`)
 
    from_braket
    to_braket
-   braket_to_qasm3
-   braket_from_qasm3
 
 """
-from qbraid.transpiler.cirq_braket.convert_braket_qasm import braket_from_qasm3, braket_to_qasm3
 from qbraid.transpiler.cirq_braket.convert_from_braket import from_braket
 from qbraid.transpiler.cirq_braket.convert_to_braket import to_braket
