+++ qbraid/transpiler/cirq_braket/tests/test_braket_qasm.py
@@ -20,11 +20,7 @@ from braket.circuits import Instruction
 from braket.circuits import gates as braket_gates
 
 from qbraid.interface import circuits_allclose
-from qbraid.transpiler.cirq_braket.convert_braket_qasm import (
-    braket_from_qasm3,
-    braket_to_qasm3,
-    from_braket,
-)
+from qbraid.transpiler.cirq_braket.convert_from_braket_qasm import from_braket
 from qbraid.transpiler.exceptions import CircuitConversionError
 
 
@@ -127,33 +123,3 @@ def test_from_braket_raises_on_unsupported_gates():
         braket_circuit.add_instruction(instr)
         with pytest.raises(CircuitConversionError):
             from_braket(braket_circuit)
-
-
-def test_braket_to_qasm3_bell_circuit():
-    """Test converting braket bell circuit to OpenQASM 3.0 string"""
-    qasm_expected = """
-OPENQASM 3.0;
-bit[2] __bits__;
-qubit[2] __qubits__;
-h __qubits__[0];
-cnot __qubits__[0], __qubits__[1];
-__bits__[0] = measure __qubits__[0];
-__bits__[1] = measure __qubits__[1];
-"""
-    bell = BKCircuit().h(0).cnot(0, 1)
-    assert qasm_expected.strip("\n") == braket_to_qasm3(bell)
-
-
-def test_braket_from_qasm3():
-    """Test converting OpenQASM 3 string to braket circuit"""
-    qasm_str = """"
-OPENQASM 3.0;
-bit[2] b;
-qubit[2] q;
-rx(0.15) q[0];
-rx(0.3) q[1];
-b[0] = measure q[0];
-b[1] = measure q[1];
-"""
-    circuit_expected = BKCircuit().rx(0, 0.15).rx(1, 0.3)
-    assert circuit_expected == braket_from_qasm3(qasm_str)
